Fix rendering of html head template
Using syntax quotes to expand the css helper function provided by
hiccup. This preserves the nesting of the persistent vector rather
than returning only the markup for the stylesheet link when the
function call is evaluated.

Not entirely sure if the private head function should actually be a defmacro -
both render the correct markup when supplying the page template to the
system handler, but I think defining it as a function is correct
because it contains edn rather than forms which should be evaluated.