Fix bug in css-watcher
I want to make the figwheel configuration a property of the system
map, but so far haven't been able to figure out how to pass the
configuration as an argument to the figwheel-system and css-watcher
components. Right now, I'm invoking it twice, which is lame, but
figwheel doesn't except a :css-dirs key in build profiles. Perhaps I
should use :watch-paths...